local/generic ssh driver for minikube
Just scp the minikube into VM or host and run:
out/minikube --v=9 start --logtostderr --vm-driver=local

* We do not need the minikube ISO, so don't download it.
* Use 127.0.0.1 as the IP
* Get the user name and the SSH id_rsa from the environment.
* Support only Linux for the moment.
* Initial code picked up from:
  https://github.com/docker/machine/blob/master/drivers/generic/generic.go

Earlier and ongoing discussions, point to overriding SSH in the
libmachine layer somehow. This PR takes the approach of of reusing
the information already available in the environment to support
the no-vm use case.

Note that it would be easy to supply the IP, id_rsa via command line
to even support a remote host/vm case if necessary.